refactor: adapt routes from profile picture to S3

+++ server/src/controllers/user/pictureController.js
@@ -1,25 +1,26 @@
-import User from "../../models/userModel.js";
+import User, { defaultPicture } from "../../models/userModel.js";
+import deleteUpload from "../../modules/multer/deleteUpload.js";
 import uploadProfilePicture from "../../modules/multer/uploads/uploadProfilePicture.js";
 import { error } from "../../utils/helpers/validations.js";
-import deleteUpload from "../../modules/multer/deleteUpload.js";
 
 export const newProfilePicture = async (req, res) => {
-    const { file } = req;
-
     try {
         await uploadProfilePicture("picture", req, res);
 
-        const { picture } = await User.findByIdAndUpdate(req.userId, {
-            picture: {
-                key: file.key,
-                source: file.location,
-            },
-        });
+        const user = await User.findById(req.userId);
+        const { picture } = user;
+
+        if (picture.key !== defaultPicture.key)
+            await deleteUpload(picture.source, picture.key);
+
+        picture.key = req.file.key;
+        picture.source = req.file.location;
+        user.save();
 
         res.send(picture);
     } catch {
         return error(
-            "Não foi possível adicionar uma foto de perfil. Tente novamente.",
+            "Não foi possível atualizar sua foto de perfil. Tente novamente.",
             500,
             res
         );
@@ -28,9 +29,21 @@ export const newProfilePicture = async (req, res) => {
 
 export const deleteProfilePicture = async (req, res) => {
     try {
-        const { picture } = await User.findById(req.userId);
+        const user = await User.findById(req.userId);
+        const { picture } = user;
+
+        if (picture.key === defaultPicture.key)
+            return error(
+                "Não é possível remover a foto de perfil padrão.",
+                400,
+                res
+            );
 
         await deleteUpload(picture.source, picture.key);
+
+        picture.key = defaultPicture.key;
+        picture.source = defaultPicture.source;
+        user.save();
 
         res.send(picture);
     } catch {
